Handle encrypted file arguments in jwplayer.
This is pretty neat. Many sites (vidbull.com is among them) have started
encrypting their file handles so that you can't read the URL to the
video file straight from the (obscured and encoded) JS source. However,
since the client eventually has to request the file, the key is stored
in the SWF. After some digging, it turns out that the key is the
`asproject.swf` plugin that is loaded. Some digging with
http://www.showmycode.com/ revealed that this file loads a file called
`obc.swf` from the same directory, which in turn does encryption and
decryption using the AS3 crypto library As3Crypto. The key is hard-coded
into the SWF in five parts (see `h1` through `h5` in `decryptor.as` in
the disassembly of `obc.swf`). Using `swfdump` from the `swftools`
package easily gets us this key, which can then be used to decrypt the
video's URL using openssl (after hex-decoding the file argument).